Fix bug with auto-ignored paths
cargo llvm-cov currently adds all the workspace members
except the one you're currently working in to the regex
passed to -ignore-filename-regex. They are added as
paths to their package directories, without a trailing
path separator. This means that if you have a workspace
with the two members 'foo' and 'food', and you enter
the directory 'food' and type 'cargo llvm-cov --html',
you will get an empty report since the path 'food' has
'foo' as a prefix, and 'foo' has been automatically added
to the ignore list.

This patch fixes this problem by adding a trailing path
separator to all the workspace members added to the
regex. 'food' does not have 'foo/' as a prefix, so the
example above should now work.