Calculate --show-missing-lines based on function regions
The `show-missing-lines-complete` case shows a situation where
`llvm-cov` reports complete line coverage, but we still reported a
single uncovered line.

Looking at the `llvm-cov` source code, the reason for this difference
was that we calculated the missing lines based on files key in the json
output, but `llvm-cov` calculates line coverage based on the (internal
equivalent of the) functions key.

So change the calculation to use a similar approach, this way both
`llvm-cov` and our code reports complete coverage for this testcase.

Also update show-missing-lines (this has missing lines) to a json which
has the "source", so it is possible to compare `llvm-cov` and our
output. Sadly when tests/fixtures/show-missing-lines.json was originally
added, I forgot to include the "source", so in its current form it was
just a change detector, which is not that useful.

Fixes <https://github.com/taiki-e/cargo-llvm-cov/issues/149>.