Improve --show-missing-lines for multiple functions in a single line
The JSON output we get from llvm-cov has stats of the amount of lines
not covered, but does not mention individual uncovered lines, we have to
calculate that.

We have calculated coverage based on iterating the list of functions,
but this did not take into account that when expanding macros, it's
normal to have multiple functions on the same line.

Fix this problem by not only recording uncovered but also covered lines,
and once we have information for all functions, then merging this
coverage info into a single list of uncovered lines, which is closer to
how llvm-cov itself calculates its own stats.

Ideally llvm-cov itself would provide this info and then we can't get it
wrong, but at least this is meant to be an incremental improvement.

Fixes <https://github.com/taiki-e/cargo-llvm-cov/issues/181>.